tools-api/pyproject.toml
[tool.poetry]
name = "tools-api"
version = "0.1.0"
description = ""
authors = ["bagdenaj <jo.bagd@posteo.de>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
psycopg2 = "^2.9.3"
uvicorn = "^0.17.6"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.3.0"
python-dotenv = "^0.20.0"
tqdm = "^4.64.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
